+++ dynamite/parameter_space.py
@@ -372,11 +372,6 @@ class ParameterGenerator(object):
             self.logger.error(text)
             raise ValueError(text)
         self.parspace_settings = parspace_settings
-        # which_chi2 = self.parspace_settings.get('which_chi2')
-        # if which_chi2 not in ['chi2', 'kinchi2']:
-        #     text = 'Unknown or missing which_chi2 setting, use chi2 or kinchi2'
-        #     self.logger.error(text)
-        #     raise ValueError(text)
         self.chi2 = self.parspace_settings.get('which_chi2')
         self.status = {}
         self.name = name